refactor: move custom validator, add constructor instead of inject method
